refactor: move ContextsStates to its own file + more unit tests Signed-off-by: Philippe Martin <phmartin@redhat.com>
