+++ packages/main/src/plugin/kubernetes/kubernetes-context-state.spec.ts
@@ -24,8 +24,8 @@ import { afterEach, beforeAll, beforeEach, describe, expect, test, vi } from 'vi
 import type { KubeContext } from '/@api/kubernetes-context.js';
 
 import type { ApiSenderType } from '../api.js';
-import type { CheckingState, ContextGeneralState, ResourceName } from './kubernetes-context-state.js';
-import { ContextsManager, ContextsStates } from './kubernetes-context-state.js';
+import type { CheckingState, ContextGeneralState, ResourceName } from './contexts-states.js';
+import { ContextsManager } from './kubernetes-context-state.js';
 
 interface InformerEvent {
   delayMs: number;
@@ -2170,49 +2170,6 @@ describe('update', async () => {
       'kubernetes-current-context-deployments-update',
       expect.anything(),
     );
-  });
-});
-describe('ContextsStates tests', () => {
-  test('hasInformer should check if informer exists for context', () => {
-    const client = new ContextsStates();
-    client.setInformers(
-      'context1',
-      new Map([['pods', new FakeInformer('context1', '/path/to/resource', 0, undefined, [], [])]]),
-    );
-    expect(client.hasInformer('context1', 'pods')).toBeTruthy();
-    expect(client.hasInformer('context1', 'deployments')).toBeFalsy();
-    expect(client.hasInformer('context2', 'pods')).toBeFalsy();
-    expect(client.hasInformer('context2', 'deployments')).toBeFalsy();
-  });
-
-  test('getContextsNames should return the names of contexts as array', () => {
-    const client = new ContextsStates();
-    client.setInformers(
-      'context1',
-      new Map([['pods', new FakeInformer('context1', '/path/to/resource', 0, undefined, [], [])]]),
-    );
-    client.setInformers(
-      'context2',
-      new Map([['pods', new FakeInformer('context2', '/path/to/resource', 0, undefined, [], [])]]),
-    );
-    expect(Array.from(client.getContextsNames())).toEqual(['context1', 'context2']);
-  });
-
-  test('isReachable', () => {
-    const client = new ContextsStates();
-    client.setInformers(
-      'context1',
-      new Map([['pods', new FakeInformer('context1', '/path/to/resource', 0, undefined, [], [])]]),
-    );
-    client.setInformers(
-      'context2',
-      new Map([['pods', new FakeInformer('context2', '/path/to/resource', 0, undefined, [], [])]]),
-    );
-    client.safeSetState('context1', state => (state.reachable = true));
-
-    expect(client.isReachable('context1')).toBeTruthy();
-    expect(client.isReachable('context2')).toBeFalsy();
-    expect(client.isReachable('context3')).toBeFalsy();
   });
 });
 describe('isContextInKubeconfig', () => {
